website-builder/ai-engine/requirements.txt
# AI Engine Service Dependencies
# Production dependencies for the AI-powered content generation engine

# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP client for service communication
httpx>=0.25.0

# Data validation and settings
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Structured logging
structlog>=23.2.0

# System monitoring
psutil>=5.9.0

# Date/time utilities
python-dateutil>=2.8.0

# JSON handling
orjson>=3.9.0

# Async support
asyncio-mqtt>=0.16.0
aiofiles>=23.2.0

# Redis for caching (optional)
redis>=5.0.0
aioredis>=2.0.0

# Database support (optional)
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0  # PostgreSQL async driver

# Background tasks (optional)
celery>=5.3.0
dramatiq>=1.15.0

# Monitoring and metrics
prometheus-client>=0.19.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing dependencies (dev)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx-mock>=0.8.0

# Code quality (dev)
black>=23.10.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0

# Development tools
pre-commit>=3.5.0